requirements.txt
Django==4.0.5
djangorestframework==3.13.1
drf-yasg==1.20.0